apps/backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Cerberus==1.3.5
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
factory_boy==3.3.1
Faker==37.4.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==10.4.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.8.0
python-dateutil==2.8.2
python-decouple==3.8
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
